add create avatar manager in dev extensions
